refactor: move to reader package
