+++ src/main/java/racingcar/reader/CarNameReader.java
@@ -1,8 +1,9 @@
-package racingcar;
+package racingcar.reader;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
+import racingcar.ConsoleDisplay;
 
 public class CarNameReader extends Reader<List<String>>{
 	public static final String INPUT_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
