+++ src/test/java/racingcar/CarNameReaderTest.java
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.MockedStatic;
+import racingcar.reader.CarNameReader;
 
 public class CarNameReaderTest {
 	MockedStatic<Console> consoleMockedStatic;
